backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytesseract==0.3.13
sqlparse==0.5.3
tesseract==0.1.3
